Refactor everything and add index.jsp with welcoming text and login/logout. This closes #2.
