+++ src/main/java/com/asegroup52/aat/Guestbook.java
@@ -15,7 +15,7 @@
  */
 
 //[START all]
-package com.example.guestbook;
+package com.asegroup52.aat;
 
 import com.googlecode.objectify.annotation.Entity;
 import com.googlecode.objectify.annotation.Id;
