+++ src/main/java/com/asegroup52/aat/RestApplication.java
@@ -1,10 +1,10 @@
-package com.example.guestbook;
+package com.asegroup52.aat;
 
 import org.restlet.Application;
 import org.restlet.Restlet;
 import org.restlet.routing.Router;
 
-public class FirstStepsApplication extends Application {
+public class RestApplication extends Application {
 
     /**
      * Creates a root Restlet that will receive all incoming calls.
