+++ src/main/webapp/WEB-INF/web2.xml
@@ -6,7 +6,7 @@
 <!-- [START standard_mappings] -->
     <servlet>
       <servlet-name>sign</servlet-name>
-      <servlet-class>com.example.guestbook.SignGuestbookServlet</servlet-class>
+      <servlet-class>com.asegroup52.aat.SignGuestbookServlet</servlet-class>
     </servlet>
 
     <servlet-mapping>
@@ -29,7 +29,7 @@
       <url-pattern>/*</url-pattern>
     </filter-mapping>
     <listener>
-      <listener-class>com.example.guestbook.OfyHelper</listener-class>
+      <listener-class>com.asegroup52.aat.OfyHelper</listener-class>
     </listener>
 <!-- [END Objectify] -->
 </web-app>
